exercises/Python/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==22.12.0
cffi==1.16.0
channels==4.0.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dj-database-url==1.2.0
Django==4.2.6
django-debug-toolbar==4.2.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
environs==9.5.0
exceptiongroup==1.1.3
iniconfig==2.0.0
isort==5.12.0
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
